+++ src/scriptengine/mod.rs
@@ -3,7 +3,9 @@
 //! Facade for all different non OS scripts.
 //!
 
+mod generic_script;
 mod rsscript;
+pub(crate) mod script_utils;
 mod shell_to_batch;
 
 #[cfg(test)]
@@ -19,6 +21,8 @@ enum EngineType {
     Rust,
     /// shell to windows batch conversion
     Shell2Batch,
+    /// Generic script runner
+    Generic,
     /// Unsupported type
     Unsupported,
 }
@@ -36,6 +40,10 @@ fn get_engine_type(task: &Task) -> EngineType {
                 } else if script_runner == "@shell" {
                     debug!("Shell to batch detected.");
                     EngineType::Shell2Batch
+                } else if task.script_extension.is_some() {
+                    // if both script runner and extension is defined, we use generic script runner
+                    debug!("Generic script detected.");
+                    EngineType::Generic
                 } else {
                     EngineType::Unsupported
                 }
@@ -58,6 +66,14 @@ pub(crate) fn invoke(task: &Task) -> bool {
         EngineType::Shell2Batch => {
             let script = task.script.as_ref().unwrap();
             shell_to_batch::execute(script);
+
+            true
+        }
+        EngineType::Generic => {
+            let script = task.script.as_ref().unwrap();
+            let runner = task.script_runner.clone().unwrap();
+            let extension = task.script_extension.clone().unwrap();
+            generic_script::execute(script, runner, extension);
 
             true
         }