+++ src/types_test.rs
@@ -45,6 +45,7 @@ fn task_new() {
     assert!(task.install_crate_args.is_none());
     assert!(task.command.is_none());
     assert!(task.disabled.is_none());
+    assert!(task.private.is_none());
     assert!(task.condition.is_none());
     assert!(task.condition_script.is_none());
     assert!(task.description.is_none());
@@ -59,6 +60,7 @@ fn task_new() {
     assert!(task.args.is_none());
     assert!(task.script.is_none());
     assert!(task.script_runner.is_none());
+    assert!(task.script_extension.is_none());
     assert!(task.run_task.is_none());
     assert!(task.dependencies.is_none());
     assert!(task.linux.is_none());
@@ -102,6 +104,7 @@ fn task_extend_both_have_misc_data() {
     base.install_crate = Some("my crate1".to_string());
     base.command = Some("test1".to_string());
     base.disabled = Some(false);
+    base.private = Some(false);
     base.script = Some(vec!["1".to_string(), "2".to_string()]);
 
     let extended = Task {
@@ -109,6 +112,7 @@ fn task_extend_both_have_misc_data() {
         command: None,
         description: None,
         disabled: Some(true),
+        private: Some(true),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -123,6 +127,7 @@ fn task_extend_both_have_misc_data() {
         args: None,
         script: None,
         script_runner: None,
+        script_extension: None,
         run_task: None,
         dependencies: None,
         linux: None,
@@ -136,6 +141,7 @@ fn task_extend_both_have_misc_data() {
     assert!(base.command.is_some());
     assert!(base.description.is_none());
     assert!(base.disabled.is_some());
+    assert!(base.private.is_some());
     assert!(base.condition.is_none());
     assert!(base.condition_script.is_none());
     assert!(base.force.is_some());
@@ -148,6 +154,7 @@ fn task_extend_both_have_misc_data() {
     assert!(base.install_crate_args.is_none());
     assert!(base.install_script.is_none());
     assert!(base.script_runner.is_none());
+    assert!(base.script_extension.is_none());
     assert!(base.run_task.is_none());
     assert!(base.args.is_none());
     assert!(base.script.is_some());
@@ -159,6 +166,7 @@ fn task_extend_both_have_misc_data() {
     assert_eq!(base.install_crate.unwrap(), "my crate2");
     assert_eq!(base.command.unwrap(), "test1");
     assert!(base.disabled.unwrap());
+    assert!(base.private.unwrap());
     assert!(base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 0);
     assert_eq!(base.alias.unwrap(), "alias2");
@@ -172,6 +180,7 @@ fn task_extend_extended_have_all_fields() {
         command: Some("test1".to_string()),
         description: None,
         disabled: Some(false),
+        private: Some(true),
         condition: None,
         condition_script: None,
         force: Some(true),
@@ -186,6 +195,7 @@ fn task_extend_extended_have_all_fields() {
         args: None,
         script: Some(vec!["1".to_string(), "2".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: None,
         linux: None,
@@ -201,6 +211,7 @@ fn task_extend_extended_have_all_fields() {
         command: Some("test2".to_string()),
         description: Some("description".to_string()),
         disabled: Some(true),
+        private: Some(false),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -220,6 +231,7 @@ fn task_extend_extended_have_all_fields() {
         args: Some(vec!["a1".to_string(), "a2".to_string()]),
         script: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
         script_runner: Some("sh2".to_string()),
+        script_extension: Some("ext2".to_string()),
         run_task: Some("task2".to_string()),
         dependencies: Some(vec!["A".to_string()]),
         linux: Some(PlatformOverrideTask {
@@ -228,6 +240,7 @@ fn task_extend_extended_have_all_fields() {
             install_crate_args: Some(vec!["c1".to_string(), "c2".to_string()]),
             command: Some("test2".to_string()),
             disabled: Some(true),
+            private: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -243,6 +256,7 @@ fn task_extend_extended_have_all_fields() {
             args: Some(vec!["a1".to_string(), "a2".to_string()]),
             script: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
             script_runner: Some("sh3".to_string()),
+            script_extension: Some("ext3".to_string()),
             run_task: Some("task3".to_string()),
             dependencies: Some(vec!["A".to_string()]),
         }),
@@ -252,6 +266,7 @@ fn task_extend_extended_have_all_fields() {
             install_crate_args: Some(vec!["c1".to_string(), "c2".to_string()]),
             command: Some("test2".to_string()),
             disabled: Some(true),
+            private: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -267,6 +282,7 @@ fn task_extend_extended_have_all_fields() {
             args: Some(vec!["a1".to_string(), "a2".to_string()]),
             script: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
             script_runner: Some("sh3".to_string()),
+            script_extension: Some("ext3".to_string()),
             run_task: Some("task3".to_string()),
             dependencies: Some(vec!["A".to_string()]),
         }),
@@ -276,6 +292,7 @@ fn task_extend_extended_have_all_fields() {
             install_crate_args: Some(vec!["c1".to_string(), "c2".to_string()]),
             command: Some("test2".to_string()),
             disabled: Some(true),
+            private: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -291,6 +308,7 @@ fn task_extend_extended_have_all_fields() {
             args: Some(vec!["a1".to_string(), "a2".to_string()]),
             script: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
             script_runner: Some("sh3".to_string()),
+            script_extension: Some("ext3".to_string()),
             run_task: Some("task3".to_string()),
             dependencies: Some(vec!["A".to_string()]),
         }),
@@ -303,6 +321,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.command.is_some());
     assert!(base.description.is_some());
     assert!(base.disabled.is_some());
+    assert!(base.private.is_some());
     assert!(base.condition.is_some());
     assert!(base.condition_script.is_some());
     assert!(base.force.is_some());
@@ -316,6 +335,7 @@ fn task_extend_extended_have_all_fields() {
     assert!(base.args.is_some());
     assert!(base.script.is_some());
     assert!(base.script_runner.is_some());
+    assert!(base.script_extension.is_some());
     assert!(base.run_task.is_some());
     assert!(base.dependencies.is_some());
     assert!(base.linux.is_some());
@@ -327,6 +347,7 @@ fn task_extend_extended_have_all_fields() {
     assert_eq!(base.command.unwrap(), "test2");
     assert_eq!(base.description.unwrap(), "description");
     assert!(base.disabled.unwrap());
+    assert!(!base.private.unwrap());
     assert_eq!(base.condition_script.unwrap().len(), 1);
     assert!(!base.force.unwrap());
     assert_eq!(base.env.unwrap().len(), 1);
@@ -339,6 +360,7 @@ fn task_extend_extended_have_all_fields() {
     assert_eq!(base.args.unwrap().len(), 2);
     assert_eq!(base.script.unwrap().len(), 3);
     assert_eq!(base.script_runner.unwrap(), "sh2");
+    assert_eq!(base.script_extension.unwrap(), "ext2");
     assert_eq!(base.run_task.unwrap(), "task2");
     assert_eq!(base.dependencies.unwrap().len(), 1);
     assert!(base.linux.unwrap().clear.unwrap());
@@ -396,6 +418,7 @@ fn task_get_normalized_task_undefined() {
         install_crate_args: None,
         command: Some("command".to_string()),
         disabled: Some(false),
+        private: Some(true),
         condition: None,
         condition_script: None,
         force: None,
@@ -405,6 +428,7 @@ fn task_get_normalized_task_undefined() {
         args: Some(vec!["1".to_string(), "2".to_string()]),
         script: Some(vec!["a".to_string(), "b".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: Some(vec!["1".to_string()]),
         description: Some("description".to_string()),
@@ -419,6 +443,7 @@ fn task_get_normalized_task_undefined() {
     assert!(normalized_task.install_crate_args.is_none());
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
+    assert!(normalized_task.private.is_some());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
     assert!(normalized_task.force.is_none());
@@ -432,6 +457,7 @@ fn task_get_normalized_task_undefined() {
     assert!(normalized_task.args.is_some());
     assert!(normalized_task.script.is_some());
     assert!(normalized_task.script_runner.is_some());
+    assert!(normalized_task.script_extension.is_some());
     assert!(normalized_task.run_task.is_some());
     assert!(normalized_task.dependencies.is_some());
     assert!(normalized_task.description.is_some());
@@ -443,6 +469,7 @@ fn task_get_normalized_task_undefined() {
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert_eq!(normalized_task.description.unwrap(), "description");
     assert!(!normalized_task.disabled.unwrap());
+    assert!(normalized_task.private.unwrap());
     assert!(!normalized_task.force.unwrap_or(false));
     assert_eq!(normalized_task.alias.unwrap(), "alias");
     assert_eq!(normalized_task.linux_alias.unwrap(), "linux");
@@ -452,6 +479,7 @@ fn task_get_normalized_task_undefined() {
     assert_eq!(normalized_task.args.unwrap().len(), 2);
     assert_eq!(normalized_task.script.unwrap().len(), 2);
     assert_eq!(normalized_task.script_runner.unwrap(), "sh1");
+    assert_eq!(normalized_task.script_extension.unwrap(), "ext1");
     assert_eq!(normalized_task.run_task.unwrap(), "task1");
     assert_eq!(normalized_task.dependencies.unwrap().len(), 1);
 }
@@ -472,6 +500,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         command: Some("command".to_string()),
         description: Some("description".to_string()),
         disabled: Some(false),
+        private: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -487,6 +516,7 @@ fn task_get_normalized_task_with_override_no_clear() {
         args: Some(vec!["1".to_string(), "2".to_string()]),
         script: Some(vec!["a".to_string(), "b".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: Some(vec!["1".to_string()]),
         linux: Some(PlatformOverrideTask {
@@ -495,6 +525,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             install_crate_args: Some(vec!["c1".to_string(), "c2".to_string(), "c3".to_string()]),
             command: Some("linux_command".to_string()),
             disabled: Some(true),
+            private: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
                 channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -515,6 +546,7 @@ fn task_get_normalized_task_with_override_no_clear() {
             args: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
             script: Some(vec!["a".to_string(), "b".to_string(), "c".to_string()]),
             script_runner: Some("sh2".to_string()),
+            script_extension: Some("ext2".to_string()),
             run_task: Some("task2".to_string()),
             dependencies: Some(vec!["1".to_string(), "2".to_string()]),
         }),
@@ -529,6 +561,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.description.is_some());
     assert!(normalized_task.disabled.is_some());
+    assert!(normalized_task.private.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -542,6 +575,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert!(normalized_task.args.is_some());
     assert!(normalized_task.script.is_some());
     assert!(normalized_task.script_runner.is_some());
+    assert!(normalized_task.script_extension.is_some());
     assert!(normalized_task.run_task.is_some());
     assert!(normalized_task.dependencies.is_some());
     assert!(normalized_task.linux.is_none());
@@ -553,6 +587,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert_eq!(normalized_task.command.unwrap(), "linux_command");
     assert_eq!(normalized_task.description.unwrap(), "description");
     assert!(normalized_task.disabled.unwrap());
+    assert!(!normalized_task.private.unwrap());
     assert_eq!(normalized_task.condition_script.unwrap().len(), 1);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -561,6 +596,7 @@ fn task_get_normalized_task_with_override_no_clear() {
     assert_eq!(normalized_task.args.unwrap().len(), 3);
     assert_eq!(normalized_task.script.unwrap().len(), 3);
     assert_eq!(normalized_task.script_runner.unwrap(), "sh2");
+    assert_eq!(normalized_task.script_extension.unwrap(), "ext2");
     assert_eq!(normalized_task.run_task.unwrap(), "task2");
     assert_eq!(normalized_task.dependencies.unwrap().len(), 2);
 
@@ -585,6 +621,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         command: Some("command".to_string()),
         description: Some("description".to_string()),
         disabled: Some(false),
+        private: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -600,6 +637,7 @@ fn task_get_normalized_task_with_override_clear_false() {
         args: Some(vec!["1".to_string(), "2".to_string()]),
         script: Some(vec!["a".to_string(), "b".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: Some(vec!["1".to_string()]),
         linux: Some(PlatformOverrideTask {
@@ -607,6 +645,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             install_crate: Some("linux_crate".to_string()),
             command: Some("linux_command".to_string()),
             disabled: Some(true),
+            private: Some(false),
             condition: Some(TaskCondition {
                 platforms: Some(vec!["linux".to_string()]),
                 channels: Some(vec![
@@ -632,6 +671,7 @@ fn task_get_normalized_task_with_override_clear_false() {
             args: Some(vec!["1".to_string(), "2".to_string(), "3".to_string()]),
             script: Some(vec!["a".to_string(), "b".to_string(), "c".to_string()]),
             script_runner: Some("sh2".to_string()),
+            script_extension: Some("ext2".to_string()),
             run_task: Some("task2".to_string()),
             dependencies: Some(vec!["1".to_string(), "2".to_string()]),
         }),
@@ -645,6 +685,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.description.is_some());
     assert!(normalized_task.disabled.is_some());
+    assert!(normalized_task.private.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -659,6 +700,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert!(normalized_task.args.is_some());
     assert!(normalized_task.script.is_some());
     assert!(normalized_task.script_runner.is_some());
+    assert!(normalized_task.script_extension.is_some());
     assert!(normalized_task.run_task.is_some());
     assert!(normalized_task.dependencies.is_some());
     assert!(normalized_task.linux.is_none());
@@ -669,6 +711,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert_eq!(normalized_task.command.unwrap(), "linux_command");
     assert_eq!(normalized_task.description.unwrap(), "description");
     assert!(normalized_task.disabled.unwrap());
+    assert!(!normalized_task.private.unwrap());
     assert_eq!(normalized_task.condition_script.unwrap().len(), 2);
     assert!(normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 1);
@@ -678,6 +721,7 @@ fn task_get_normalized_task_with_override_clear_false() {
     assert_eq!(normalized_task.args.unwrap().len(), 3);
     assert_eq!(normalized_task.script.unwrap().len(), 3);
     assert_eq!(normalized_task.script_runner.unwrap(), "sh2");
+    assert_eq!(normalized_task.script_extension.unwrap(), "ext2");
     assert_eq!(normalized_task.run_task.unwrap(), "task2");
     assert_eq!(normalized_task.dependencies.unwrap().len(), 2);
 
@@ -698,6 +742,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         install_crate_args: Some(vec!["c1".to_string(), "c2".to_string()]),
         command: Some("command".to_string()),
         disabled: Some(false),
+        private: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -713,6 +758,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
         args: Some(vec!["1".to_string(), "2".to_string()]),
         script: Some(vec!["a".to_string(), "b".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: Some(vec!["1".to_string()]),
         description: None,
@@ -722,6 +768,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             install_crate_args: None,
             command: None,
             disabled: None,
+            private: None,
             condition: None,
             condition_script: None,
             force: None,
@@ -731,6 +778,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
             args: None,
             script: None,
             script_runner: None,
+            script_extension: None,
             run_task: None,
             dependencies: None,
         }),
@@ -744,6 +792,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert!(normalized_task.install_crate_args.is_some());
     assert!(normalized_task.command.is_some());
     assert!(normalized_task.disabled.is_some());
+    assert!(normalized_task.private.is_some());
     assert!(normalized_task.condition.is_some());
     assert!(normalized_task.condition_script.is_some());
     assert!(normalized_task.force.is_some());
@@ -757,6 +806,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert!(normalized_task.args.is_some());
     assert!(normalized_task.script.is_some());
     assert!(normalized_task.script_runner.is_some());
+    assert!(normalized_task.script_extension.is_some());
     assert!(normalized_task.run_task.is_some());
     assert!(normalized_task.dependencies.is_some());
     assert!(normalized_task.description.is_none());
@@ -767,6 +817,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.install_crate.unwrap(), "install_crate");
     assert_eq!(normalized_task.command.unwrap(), "command");
     assert!(!normalized_task.disabled.unwrap());
+    assert!(normalized_task.private.unwrap());
     assert!(!normalized_task.force.unwrap());
     assert_eq!(normalized_task.env.unwrap().len(), 0);
     assert_eq!(normalized_task.cwd.unwrap(), "cwd".to_string());
@@ -775,6 +826,7 @@ fn task_get_normalized_task_with_override_clear_false_partial_override() {
     assert_eq!(normalized_task.args.unwrap().len(), 2);
     assert_eq!(normalized_task.script.unwrap().len(), 2);
     assert_eq!(normalized_task.script_runner.unwrap(), "sh1");
+    assert_eq!(normalized_task.script_extension.unwrap(), "ext1");
     assert_eq!(normalized_task.run_task.unwrap(), "task1");
     assert_eq!(normalized_task.dependencies.unwrap().len(), 1);
 }
@@ -791,6 +843,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         install_crate_args: Some(vec!["c1".to_string(), "c2".to_string()]),
         command: Some("command".to_string()),
         disabled: Some(false),
+        private: Some(true),
         condition: Some(TaskCondition {
             platforms: Some(vec!["linux".to_string(), "mac".to_string()]),
             channels: Some(vec!["nightly".to_string(), "stable".to_string()]),
@@ -806,6 +859,7 @@ fn task_get_normalized_task_with_override_clear_true() {
         args: Some(vec!["1".to_string(), "2".to_string()]),
         script: Some(vec!["a".to_string(), "b".to_string()]),
         script_runner: Some("sh1".to_string()),
+        script_extension: Some("ext1".to_string()),
         run_task: Some("task1".to_string()),
         dependencies: Some(vec!["1".to_string()]),
         description: Some("description".to_string()),
@@ -815,6 +869,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             install_crate_args: None,
             command: None,
             disabled: None,
+            private: None,
             condition: None,
             condition_script: None,
             force: None,
@@ -824,6 +879,7 @@ fn task_get_normalized_task_with_override_clear_true() {
             args: None,
             script: None,
             script_runner: None,
+            script_extension: None,
             run_task: None,
             dependencies: None,
         }),
@@ -837,6 +893,7 @@ fn task_get_normalized_task_with_override_clear_true() {
     assert!(normalized_task.install_crate_args.is_none());
     assert!(normalized_task.command.is_none());
     assert!(normalized_task.disabled.is_none());
+    assert!(normalized_task.private.is_none());
     assert!(normalized_task.condition.is_none());
     assert!(normalized_task.condition_script.is_none());
     assert!(normalized_task.force.is_none());
@@ -850,6 +907,7 @@ fn task_get_normalized_task_with_override_clear_true() {
     assert!(normalized_task.args.is_none());
     assert!(normalized_task.script.is_none());
     assert!(normalized_task.script_runner.is_none());
+    assert!(normalized_task.script_extension.is_none());
     assert!(normalized_task.run_task.is_none());
     assert!(normalized_task.dependencies.is_none());
     assert!(normalized_task.description.is_some());